Server/requirements.txt
flask
flask-cors
googlemaps
python-dotenv
pymongo[srv]
bcrypt
uuid